[DashboardLayout] Update stories
